move LocationReportingTask out of wifi package
